fix(ci): align node worker bridge checks
